refactor: move UserContext to a separate file and update imports in components
