+++ frontend/src/components/Navbar/UserInfoName/UserInfoName.tsx
@@ -1,6 +1,6 @@
 import { useContext } from "react";
 import { Navbar } from "react-bootstrap";
-import { UserContext } from "../../../App";
+import { UserContext } from "../../../stores/UserContext";
 
 export function UserInfoName() {
   const { userInfo } = useContext(UserContext);
