+++ frontend/src/stores/UserContext.ts
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import type { UserInfo } from "../types/github/User";
+
+export const UserContext = createContext<UserInfo | null>(null);
